requirements.txt
asgiref==3.7.2
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
contourpy==1.1.0
crispy==0.7.4
crispy-bootstrap4==2022.1
cryptography==41.0.5
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.58.1
django-braces==1.15.0
django-crispy-forms==2.0
django-model-utils==4.3.1
django-pwa==1.1.0
django-star-ratings==0.9.2
django-storages==1.14.2
django-widget-tweaks==1.5.0
fabio==2023.6.0
fonttools==4.42.1
h5py==3.9.0
idna==3.4
isodate==0.6.1
kiwisolver==1.4.5
matplotlib==3.7.2
mysql==0.0.3
mysqlclient==2.2.0
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.1.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
silx==1.1.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
